requirements.txt
# Untuk Bot Telegram
python-telegram-bot

lxml
aiolimiter~=1.2.1
anyio~=4.9.0
apngasm_python~=1.3.2
av>=13.1.0
beautifulsoup4~=4.13.4
cryptg~=0.5.0.post0
rookiepy~=0.5.6
httpx~=0.28.1
imagequant~=1.1.4
memory-tempfile~=2.2.3
mergedeep~=1.3.4
numpy>=1.22.4
Pillow~=11.2.1
pyoxipng~=9.1.0
python-telegram-bot~=22.2
psutil~=7.0.0
PyMemoryEditor~=1.5.22
requests~=2.32.4
rlottie_python~=1.3.7
signalstickers-client-fork-laggykiller~=3.3.0.post2
socksio~=1.0.0
telethon~=1.40.0
tqdm~=4.67.1
ttkbootstrap-fork-laggykiller~=1.5.1
websocket_client~=1.8.0

# Untuk Fitur Bot Status
psutil